Add files via upload
Livejournal and Twitter table of topic dynamics (LDA topics within larger word2vec-based topics) to trace changes in the topic tokens across the weeks. '_com' - tokens shared by similar topics; '_lost' - tokens that did not appear in a given week; '_added' - new tokens that formed part of the topic representation in a given week; '_rat_1t_users' - ratio of the narrowly focused audience in a given week, 1-topic users; '_rat_users' - the ratio of the number of users in a given week that contributed to this discussion.